Rename utility package to util
