+++ src/main/java/net/smartcosmos/edge/things/rest/request/ThingRequestFactory.java
@@ -12,7 +12,7 @@ import net.smartcosmos.edge.things.config.SmartCosmosEdgeThingsProperties;
 import net.smartcosmos.edge.things.domain.things.RestThingCreate;
 import net.smartcosmos.edge.things.domain.things.RestThingUpdate;
 
-import static net.smartcosmos.edge.things.utility.UrlEncodingUtil.encode;
+import static net.smartcosmos.edge.things.util.UrlEncodingUtil.encode;
 
 /**
  * Utility component for creating requests for the Thing service.
