+++ src/main/java/net/smartcosmos/edge/things/service/CreateThingEdgeServiceDefault.java
@@ -18,7 +18,7 @@ import net.smartcosmos.edge.things.service.metadata.CreateMetadataRestService;
 import net.smartcosmos.edge.things.service.things.CreateThingRestService;
 import net.smartcosmos.security.user.SmartCosmosUser;
 
-import static net.smartcosmos.edge.things.utility.ResponseBuilderUtility.buildForwardingResponse;
+import static net.smartcosmos.edge.things.util.ResponseBuilderUtility.buildForwardingResponse;
 
 /**
  * Default implementation for {@link net.smartcosmos.edge.things.service.CreateThingEdgeService}
