+++ src/main/java/net/smartcosmos/edge/things/service/DeleteThingEdgeServiceDefault.java
@@ -14,7 +14,7 @@ import net.smartcosmos.edge.things.service.metadata.DeleteMetadataRestService;
 import net.smartcosmos.edge.things.service.things.DeleteThingRestService;
 import net.smartcosmos.security.user.SmartCosmosUser;
 
-import static net.smartcosmos.edge.things.utility.ResponseBuilderUtility.buildForwardingResponse;
+import static net.smartcosmos.edge.things.util.ResponseBuilderUtility.buildForwardingResponse;
 
 /**
  * Default implementation for {@link net.smartcosmos.edge.things.service.DeleteThingEdgeService}
