+++ src/main/java/net/smartcosmos/edge/things/service/UpdateThingEdgeServiceDefault.java
@@ -17,8 +17,8 @@ import net.smartcosmos.edge.things.service.metadata.UpsertMetadataRestService;
 import net.smartcosmos.edge.things.service.things.UpdateThingRestService;
 import net.smartcosmos.security.user.SmartCosmosUser;
 
-import static net.smartcosmos.edge.things.utility.ResponseBuilderUtility.buildBadRequestResponse;
-import static net.smartcosmos.edge.things.utility.ResponseBuilderUtility.buildForwardingResponse;
+import static net.smartcosmos.edge.things.util.ResponseBuilderUtility.buildBadRequestResponse;
+import static net.smartcosmos.edge.things.util.ResponseBuilderUtility.buildForwardingResponse;
 
 /**
  * Default implementation for {@link net.smartcosmos.edge.things.service.UpdateThingEdgeService}
