+++ src/main/java/net/smartcosmos/edge/things/service/event/EventSendingService.java
@@ -1,6 +1,6 @@
 package net.smartcosmos.edge.things.service.event;
 
-import net.smartcosmos.edge.things.utility.ThingEdgeEventType;
+import net.smartcosmos.edge.things.util.ThingEdgeEventType;
 import net.smartcosmos.security.user.SmartCosmosUser;
 
 /**
