+++ src/main/java/net/smartcosmos/edge/things/util/ResponseBuilderUtility.java
@@ -1,4 +1,4 @@
-package net.smartcosmos.edge.things.utility;
+package net.smartcosmos.edge.things.util;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
