+++ src/main/java/net/smartcosmos/edge/things/util/ThingEdgeEventType.java
@@ -1,4 +1,4 @@
-package net.smartcosmos.edge.things.utility;
+package net.smartcosmos.edge.things.util;
 
 import org.apache.commons.lang.StringUtils;
